refactor db connection and add error handling and db queries to provision as well

+++ deletionhandler.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	apiv1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"log"
 	"net/http"
 )
 
@@ -13,16 +14,21 @@ func DeletionHandler(w http.ResponseWriter, r *http.Request) {
 	var tm Trader
 	dr := DeletionRequest{}
 	_ = json.NewDecoder(r.Body).Decode(&dr)
-	db.Where("name = ?", dr.Name).First(&tm)
+	db.instance.Where("name = ?", dr.Name).First(&tm)
+	dbResult := db.instance.Delete(&tm)
 
-	fmt.Println("Deleting deployment...")
-	deletePolicy := metav1.DeletePropagationForeground
-	deploymentsClient := kubernetesClientSet.AppsV1().Deployments(apiv1.NamespaceDefault)
-	if err := deploymentsClient.Delete(context.TODO(), "trader-"+tm.TraderId, metav1.DeleteOptions{
-		PropagationPolicy: &deletePolicy,
-	}); err != nil {
-		panic(err)
+	if dbResult.Error != nil {
+		log.Printf("Could not delete trader for user %s, with ID of %s", tm.Name, tm.TraderId)
+		http.Error(w, fmt.Sprintf("Could not delete trader for user %s, with ID of %s", tm.Name, tm.TraderId), http.StatusBadRequest)
+	} else {
+		log.Println("Deleting deployment...")
+		deletePolicy := metav1.DeletePropagationForeground
+		deploymentsClient := kubernetesClientSet.AppsV1().Deployments(apiv1.NamespaceDefault)
+		if err := deploymentsClient.Delete(context.TODO(), "trader-"+tm.TraderId, metav1.DeleteOptions{
+			PropagationPolicy: &deletePolicy,
+		}); err != nil {
+			panic(err)
+		}
+		log.Println("Deleted deployment.")
 	}
-	fmt.Println("Deleted deployment.")
-	db.Delete(&tm)
 }
